OnnxPipleline-main/requirements.txt
numpy==1.21.6
paho-mqtt==2.1.0
onnxruntime==1.17.3